Use float for angle computations

+++ src/entities/Enemy.cpp
@@ -312,7 +312,7 @@ void Enemy::beAbox()
 		Fixed r = itofix(1);
 		if (internal[0] & 1)
 		{
-			Fixed a = (Fixed)(atan2((float)dy + r, (float)dx) * 128 / M_PI);
+			Fixed a = (Fixed)(atan2f((float)(dy + r), (float)dx) * 128.f / M_PI);
 			if (!(internal[2] & 1))
 				for (int i = 0; i < 3; i++)
 					Level::bArray->add(x + itofix(i - 1) * 10, y + itofix(24), a, r, static_cast<int>(LUTs::BaseImageId::ENEMY_BULLET_1_LIGHT), polarity, true, getCamRelation());
@@ -321,7 +321,7 @@ void Enemy::beAbox()
 		}
 		if (internal[0] & 2)
 		{
-			Fixed a = 128 - (Fixed)(asin((float)dy / (float)r) * 128 / M_PI);
+			Fixed a = 128 - (Fixed)(asinf((float)dy / (float)r) * 128.f / M_PI);
 			if (!(internal[2] & 2))
 				for (int i = 0; i < 3; i++)
 					Level::bArray->add(x - itofix(24), y + itofix(i - 1) * 10, a, r, static_cast<int>(LUTs::BaseImageId::ENEMY_BULLET_1_LIGHT), polarity, true, getCamRelation());
@@ -330,7 +330,7 @@ void Enemy::beAbox()
 		}
 		if (internal[0] & 4)
 		{
-			Fixed a = (Fixed)(asin((float)dy / (float)r) * 128 / M_PI);
+			Fixed a = (Fixed)(asinf((float)dy / (float)r) * 128.f / M_PI);
 
 			if (!(internal[2] & 4))
 				for (int i = 0; i < 3; i++)
@@ -340,7 +340,7 @@ void Enemy::beAbox()
 		}
 		if (internal[0] & 8)
 		{
-			Fixed a = (Fixed)(atan2((float)dy - r, (float)dx) * 128 / M_PI);
+			Fixed a = (Fixed)(atan2f((float)(dy - r), (float)dx) * 128.f / M_PI);
 			if (!(internal[2] & 8))
 				for (int i = 0; i < 3; i++)
 					Level::bArray->add(x + itofix(i - 1) * 10, y - itofix(24), a, r, static_cast<int>(LUTs::BaseImageId::ENEMY_BULLET_1_LIGHT), polarity, true, getCamRelation());
